+++ android/src/main/java/expo/modules/superwallexpo/json/PaywallResult.kt
@@ -3,18 +3,20 @@ package expo.modules.superwallexpo.json
 import com.superwall.sdk.paywall.presentation.internal.state.PaywallResult
 
 fun PaywallResult.toJson(): Map<String, Any> {
-        return mutableMapOf<String, Any>().apply {
-            when (this) {
-                is PaywallResult.Purchased -> {
-                    this["type"] = "purchased"
-                    this["productId"] = productId
-                }
-                is PaywallResult.Declined -> {
-                    this["type"] = "declined"
-                }
-                is PaywallResult.Restored -> {
-                    this["type"] = "restored"
-                }
-            }
+    val map = mutableMapOf<String, Any>()
+    
+    when (this) {  
+        is PaywallResult.Purchased -> {
+            map["type"] = "purchased"
+            map["productId"] = productId
+        }
+        is PaywallResult.Declined -> {
+            map["type"] = "declined"
+        }
+        is PaywallResult.Restored -> {
+            map["type"] = "restored"
         }
     }
+    
+    return map
+}