+++ app.plugin.js
@@ -1,2 +1,0 @@
-// This file configures the entry file for your plugin.
-module.exports = require("./plugin/build");