+++ ios/Json/ProductStore+Json.swift
@@ -7,7 +7,14 @@ extension ProductStore {
       return "APP_STORE"
     case .stripe:
       return "STRIPE"
+    case .paddle:
+      return "PADDLE"
+    case .superwall:
+      return "SUPERWALL"
+    case .playStore:
+      return "PLAY_STORE"
+    case .other:
+      return "OTHER"
     }
-
   }
 }